Added two tests for wrapping a single line and for two lines
The first testcase is quite trivial, but it is a nice preparation for implementing a recursive function.
If there is only one line, you can just return an array containing this line.

The second test pretty much covers the first "working" implementation (beside the fact that there is no word wrapping)